some fixes on hr.

+++ Framework.AuthServer/Dtos/JobService/Output/TimeToHireReportOutput.cs
@@ -5,7 +5,7 @@ public class TimeToHireReportOutput
     public List<TimeToHireData> Data { get; set; } = [];
     public decimal AverageTimeToHire { get; set; }
     public int TotalHired { get; set; }
-    public int TotalApplications { get; set; }
+    public long TotalApplications { get; set; }
 }
 
 public class TimeToHireData
